+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `HttpAuthTasks` docs.
 - Update `IsPathTooLong` to equal or greater than `260`.
 - Update `SignExtension.Sign` to sign NuGet or files.
+- Update `SignProject` to sign files using `Azure Key Vault` if available.
 ### Tests
 - Update `NuGetExtensionTests`
 - Add `AzureKeyVaultConfigTests`